+++ src/main/scala/sc4pac/service.scala
@@ -5,9 +5,9 @@ package service
 
 trait FileSystem {
 
-  private lazy val projDirs = dev.dirs.ProjectDirectories.from("", cli.BuildInfo.organization, cli.BuildInfo.name)  // qualifier, organization, application
-  def projectConfigDir: String = projDirs.configDir
-  def projectCacheDir: String = projDirs.cacheDir
+  private lazy val projDirs = scala.util.Try(dev.dirs.ProjectDirectories.from("", cli.BuildInfo.organization, cli.BuildInfo.name))  // qualifier, organization, application
+  def projectConfigDir: String = projDirs.get.configDir
+  def projectCacheDir: String = projDirs.get.cacheDir
 
   def readEnvVar(name: String): Option[String] =
     Option(System.getenv(name)).filter(_.nonEmpty)
@@ -23,6 +23,10 @@ trait FileSystem {
     lazy val sc4pacProfilesDir: Option[String] = readEnvVar("SC4PAC_PROFILES_DIR")
     /** Personal access token for authenticating to Simtropolis. */
     lazy val simtropolisToken: Option[String] = readEnvVar("SC4PAC_SIMTROPOLIS_TOKEN")
+    /** Used in case ProjectDirectories fails to find %appdata% directory. */
+    lazy val appDataWindows: Option[String] = readEnvVar("APPDATA")  // TODO use only on Windows
+    /** Used in case ProjectDirectories fails to find %localappdata% directory. */
+    lazy val localAppDataWindows: Option[String] = readEnvVar("LOCALAPPDATA")  // TODO use only on Windows
   }
 
   def injectErrorInTest: zio.Task[Unit] = zio.ZIO.unit